+++ examples/oauth2/src/SecurityService.scala
@@ -4,7 +4,8 @@ import scala.jdk.OptionConverters.*
 import org.pac4j.core.config.Config
 import org.pac4j.core.util.FindBest
 import org.pac4j.undertow.context.{UndertowSessionStore, UndertowWebContext}
-import ba.sake.sharaf.UndertowSharafRequest
+import ba.sake.sharaf.undertow.UndertowSharafRequest
+
 
 class SecurityService(config: Config) {
