+++ examples/user-pass-form/src/userpassform/AppRoutes.scala
@@ -2,8 +2,7 @@ package userpassform
 
 import scalatags.Text.all.*
 import ba.sake.sharaf.*
-import ba.sake.sharaf.routing.*
-import ba.sake.sharaf.undertow.UndertowSharafRoutes
+import ba.sake.sharaf.undertow.{*, given}
 
 class AppRoutes(callbackUrl: String, securityService: SecurityService) {
   val routes = UndertowSharafRoutes {
