+++ sharaf-undertow/src/ba/sake/sharaf/undertow/handlers/ExceptionHandler.scala
@@ -1,9 +1,10 @@
-package ba.sake.sharaf.handlers
+package ba.sake.sharaf.undertow.handlers
 
 import scala.util.control.NonFatal
 import io.undertow.server.HttpHandler
 import io.undertow.server.HttpServerExchange
 import ba.sake.sharaf.*
+import ba.sake.sharaf.undertow.*
 import ba.sake.sharaf.exceptions.ExceptionMapper
 
 final class ExceptionHandler private (next: HttpHandler, exceptionMapper: ExceptionMapper) extends HttpHandler {
