+++ sharaf-undertow/src/ba/sake/sharaf/undertow/handlers/SharafHandler.scala
@@ -1,4 +1,4 @@
-package ba.sake.sharaf.handlers
+package ba.sake.sharaf.undertow.handlers
 
 import io.undertow.server.HttpHandler
 import io.undertow.server.HttpServerExchange
