+++ sharaf-undertow/test/src/ba/sake/sharaf/undertow/CookiesTest.scala
@@ -1,5 +1,6 @@
-package ba.sake.sharaf
+package ba.sake.sharaf.undertow
 
+import ba.sake.sharaf.*
 import ba.sake.sharaf.undertow.*
 
 class CookiesTest extends munit.FunSuite {
