+++ sharaf-undertow/test/src/ba/sake/sharaf/undertow/FormParsingTest.scala
@@ -1,4 +1,4 @@
-package ba.sake.sharaf
+package ba.sake.sharaf.undertow
 
 import scala.collection.immutable.SeqMap
 import io.undertow.server.handlers.form.FormData as UFormData
