+++ sharaf-undertow/test/src/ba/sake/sharaf/undertow/HeadersTest.scala
@@ -1,6 +1,6 @@
-package ba.sake.sharaf
+package ba.sake.sharaf.undertow
 
-import io.undertow.Undertow
+import ba.sake.sharaf.*
 import ba.sake.sharaf.undertow.*
 
 class HeadersTest extends munit.FunSuite {
