+++ sharaf-undertow/test/src/ba/sake/sharaf/undertow/handlers/ErrorHandlerTest.scala
@@ -1,4 +1,4 @@
-package ba.sake.sharaf.handlers
+package ba.sake.sharaf.undertow.handlers
 
 import ba.sake.formson.FormDataRW
 import ba.sake.querson.QueryStringRW
